+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Dtos/Video/SearchVideosByUpIdFullDto.cs
@@ -31,8 +31,7 @@ public class SearchVideosByUpIdDto : IWrid
 
     public string order_avoided { get; set; } = "true";
 
-    public string w_rid { get; set; } //= "b280d7bc02d653ffa06d874a74e5bfd9"; //todo：应该是md5
+    public string? w_rid { get; set; }
 
-    //public long wts { get; set; } = TimeStampHelper.DateTimeToTimeStamp(DateTime.Now);
-    public long wts { get; set; } //= 1684857205; //todo
+    public long wts { get; set; }
 }