+++ src/Ray.BiliBiliTool.Agent/QingLong/Dtos/TokenResponse.cs
@@ -1,0 +1,8 @@
+namespace Ray.BiliBiliTool.Agent.QingLong.Dtos;
+
+public class TokenResponse
+{
+    public required string token { get; set; }
+
+    public required string token_type { get; set; }
+}