+++ src/Ray.BiliBiliTool.Config/Options/QingLongOptions.cs
@@ -1,0 +1,8 @@
+namespace Ray.BiliBiliTool.Config.Options;
+
+public class QingLongOptions
+{
+    public string? ClientId { get; set; }
+
+    public string? ClientSecret { get; set; }
+}