+++ src/Ray.BiliBiliTool.Infrastructure/Extensions/ICollectionExtensions.cs
@@ -1,23 +1,15 @@
-﻿using System.Linq;
-
-namespace System.Collections.Generic;
+﻿namespace Ray.BiliBiliTool.Infrastructure.Extensions;
 
 public static class ICollectionExtensions
 {
     public static void AddIfNotExist<T>(this ICollection<T> source, T add)
     {
-        if (source == null)
-            return;
-
-        if (!source.Any(x => x.Equals(add)))
+        if (!source.Any(x => x != null && x.Equals(add)))
             source.Add(add);
     }
 
     public static void AddIfNotExist<T>(this ICollection<T> source, T add, Func<T, bool> exist)
     {
-        if (source == null)
-            return;
-
         if (!source.Any(exist))
             source.Add(add);
     }